+++ core/src/BootstrapDataSpec.hpp
@@ -81,7 +81,7 @@ namespace models::stats {
 
     for (const G& group : data.classes) {
       const int group_size = data.group_size(group);
-      const int group_sample_size = std::max(1.0, std::round(group_size / (double)data_size * size));
+      const int group_sample_size = std::max(1, (int)std::round(group_size / (float)data_size * size));
 
       for (int i = 0; i < group_sample_size; i++) {
         const Uniform unif(data.group_start(group), data.group_end(group));
